The service UserService and BookService covered

+++ src/main/java/edu/eci/cvds/sistemabiblioteca/repository/BookRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer> {}
+public interface BookRepository extends JpaRepository<Book, Long> {}
